Remove duplicated protected params definitions
Use "support/stubs/strong_parameters" instead.
