+++ activerecord/test/cases/finder_test.rb
@@ -226,14 +226,14 @@ class FinderTest < ActiveRecord::TestCase
   end
 
   def test_exists_with_strong_parameters
-    assert_equal false, Subscriber.exists?(Parameters.new(nick: "foo").permit!)
+    assert_equal false, Subscriber.exists?(ProtectedParams.new(nick: "foo").permit!)
 
     Subscriber.create!(nick: "foo")
 
-    assert_equal true, Subscriber.exists?(Parameters.new(nick: "foo").permit!)
+    assert_equal true, Subscriber.exists?(ProtectedParams.new(nick: "foo").permit!)
 
     assert_raises(ActiveModel::ForbiddenAttributesError) do
-      Subscriber.exists?(Parameters.new(nick: "foo"))
+      Subscriber.exists?(ProtectedParams.new(nick: "foo"))
     end
   end
