+++ activerecord/test/cases/forbidden_attributes_protection_test.rb
@@ -1,48 +1,12 @@
 # frozen_string_literal: true
 
 require "cases/helper"
-require "active_support/core_ext/hash/indifferent_access"
-
 require "models/company"
 require "models/person"
 require "models/ship"
 require "models/ship_part"
 require "models/treasure"
-
-class ProtectedParams
-  attr_accessor :permitted
-  alias :permitted? :permitted
-
-  delegate :keys, :key?, :has_key?, :empty?, to: :@parameters
-
-  def initialize(attributes)
-    @parameters = attributes.with_indifferent_access
-    @permitted = false
-  end
-
-  def permit!
-    @permitted = true
-    self
-  end
-
-  def [](key)
-    @parameters[key]
-  end
-
-  def to_h
-    @parameters
-  end
-
-  def stringify_keys
-    dup
-  end
-
-  def dup
-    super.tap do |duplicate|
-      duplicate.instance_variable_set :@permitted, @permitted
-    end
-  end
-end
+require "support/stubs/strong_parameters"
 
 class ForbiddenAttributesProtectionTest < ActiveRecord::TestCase
   def test_forbidden_attributes_cannot_be_used_for_mass_assignment
