+++ activerecord/test/support/stubs/strong_parameters.rb
@@ -1,6 +1,10 @@
 # frozen_string_literal: true
 
-class Parameters
+require "active_support/core_ext/hash/indifferent_access"
+
+class ProtectedParams
+  delegate :keys, :key?, :has_key?, :empty?, to: :@parameters
+
   def initialize(parameters = {})
     @parameters = parameters.with_indifferent_access
     @permitted = false
@@ -15,7 +19,22 @@ class Parameters
     self
   end
 
+  def [](key)
+    @parameters[key]
+  end
+
   def to_h
     @parameters.to_h
+  end
+  alias to_unsafe_h to_h
+
+  def stringify_keys
+    dup
+  end
+
+  def dup
+    super.tap do |duplicate|
+      duplicate.instance_variable_set :@permitted, @permitted
+    end
   end
 end
